Added Redis

+++ SimpleAPI/Controllers/InformationsController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using SimpleAPI.BL.Cache;
 using SimpleAPI.BL.DTO.Information;
 using SimpleAPI.Core.Entities;
 using SimpleAPI.Core.Repository;
@@ -7,21 +8,28 @@ namespace SimpleAPI.Controllers
 {
     [Route("api/[controller]/[action]")]
     [ApiController]
-    public class InformationsController(IGenericRepository<Information> _repo) : ControllerBase
+    public class InformationsController(IGenericRepository<Information> _repo,ICacheService _cache) : ControllerBase
     {
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            string cacheKey = "Info_GetAll";
+            var cachedData = await _cache.GetAsync<List<Information>>(cacheKey);
             var data = await _repo.GetAllAsync();
             if (data is null || !data.Any()) return BadRequest();
+            await _cache.SetAsync(cacheKey, data);
             return Ok(data);
         }
 
         [HttpGet]
         public async Task<IActionResult> GetById(int id)
         {
+            if (id == 0) return BadRequest();
+            string cacheKey = $"Info_GetById_{id}";
+            var cachedItem = await _cache.GetAsync<Information>(cacheKey);
             var data = await _repo.GetByIdAsync(id);
             if (data is null) return NotFound();
+            await _cache.SetAsync(cacheKey, data);
             return Ok(data);
         }
 
@@ -37,12 +45,14 @@ namespace SimpleAPI.Controllers
                 ProductName = dto.ProductName,
             };
             await _repo.CreateAsync(inf);
+            await _cache.RemoveAsync("Info_GetAll");
             return Created($"/api/informations/{inf.Id}", inf);
         }
 
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(int id, InformationUpdateDTO dto)
         {
+            if (id == 0) return BadRequest();
             if (dto == null || !ModelState.IsValid) return BadRequest();
             var data = await _repo.GetByIdAsync(id);
             if (data is null) return BadRequest();
@@ -51,15 +61,20 @@ namespace SimpleAPI.Controllers
             data.Price = dto.Price;
             data.Description = dto.Description;
             await _repo.UpdateAsync(data);
+            await _cache.RemoveAsync("Info_GetAll");
+            await _cache.RemoveAsync($"Info_GetById_{id}");
             return NoContent();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (id == 0) return BadRequest();
             var data = await _repo.GetByIdAsync(id);
             if (data is null) return NotFound();
             await _repo.DeleteAsync(id);
+            await _cache.RemoveAsync("Info_GetAll");
+            await _cache.RemoveAsync($"Info_GetById_{id}");
             return NoContent();
         }
 
@@ -81,8 +96,15 @@ namespace SimpleAPI.Controllers
                 data.Add(inf);
             }
             await _repo.BulkInsertAsync(data);
-            return Created("Created",data);
+            return Created("Created", data);
         }
-
+        [HttpGet]
+        public async Task<IActionResult> Search(string query)
+        {
+            if (string.IsNullOrWhiteSpace(query)) return BadRequest();
+            var data = _repo.Search(x => x.ProductName.ToLower().Contains(query) || x.Description.ToLower().Contains(query) || x.Price.ToString().Contains(query) || x.CategoryId.ToString().Contains(query));
+            if (data is null) return BadRequest();
+            return Ok(data);
+        }
     }
 }
